pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcposprint"
version = "1.0.0"
description = "MCP server for thermal printer task card generation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MCPOSprint Contributors"}
]
keywords = ["mcp", "server", "thermal", "printer", "task", "cards", "escpos", "usb"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "Topic :: Printing",
]
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "qrcode[pil]>=7.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "python-escpos>=3.0.0",
    "pyusb>=1.3.1",
    "mcp[cli]>=1.10.1",
]

[project.urls]
Homepage = "https://github.com/username/mcposprint"
Repository = "https://github.com/username/mcposprint"
Issues = "https://github.com/username/mcposprint/issues"
Documentation = "https://github.com/username/mcposprint#readme"

[project.scripts]
mcposprint = "server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mcposprint"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
